diary-agent/backend/pyproject.toml
[project]
name = "vplayer-backend"
version = "0.1.0"
description = "VPlayer backend architecture"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "fastapi[standard]>=0.135.1",
    "google-genai>=1.66.0",
    "langchain-google-genai>=2.0.0",
    "langgraph>=1.0.10",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
]
